refactor: move offsite methods into trait and interface
